# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Hotel AI OS Core API
  version: 0.2.3-pilot.6
  description: >-
    PRODUCT-V1.2 / TECH-V0.2 Sprint 2.1 UAT-candidate contract. Production requests use a
    trusted JWT/SSO bearer token; roles, permissions, organization scope and
    active position assignments are resolved by the server. X-Tenant-Id and
    X-Actor-Id are accepted only when DEV_HEADER_AUTH_ENABLED=true for local
    integration. Client-supplied role or organization-scope headers are never
    trusted.
servers:
  - url: http://localhost:8080
security:
  - BearerAuth: []
  - TenantHeader: []
    ActorHeader: []
...
paths:
...
  /api/v1/work-packages:
    get:
      tags: [Work Package]
      operationId: listWorkPackages
      summary: List visible work package definitions
      responses:
        '200': { $ref: '#/components/responses/ListResponse' }
    post:
      tags: [Work Package]
      operationId: createWorkPackage
      summary: Create a work package definition
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateWorkPackage' }
      responses:
        '201': { $ref: '#/components/responses/ObjectResponse' }
  /api/v1/work-packages/{workPackageId}:
    get:
      tags: [Work Package]
      operationId: getWorkPackage
      summary: Get a work package with versions
      parameters:
        - $ref: '#/components/parameters/WorkPackageId'
      responses:
        '200': { $ref: '#/components/responses/ObjectResponse' }
  /api/v1/work-packages/{workPackageId}/versions:
    post:
      tags: [Work Package]
      operationId: createWorkPackageVersion
      summary: Create a draft work package version
      parameters:
        - $ref: '#/components/parameters/WorkPackageId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateWorkPackageVersion' }
      responses:
        '201': { $ref: '#/components/responses/ObjectResponse' }
  /api/v1/work-packages/{workPackageId}/versions/{versionId}:
    put:
      tags: [Work Package]
      operationId: updateWorkPackageVersion
      summary: Replace draft version scopes and items
      parameters:
        - $ref: '#/components/parameters/WorkPackageId'
        - $ref: '#/components/parameters/VersionId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateWorkPackageVersion' }
      responses:
        '200': { $ref: '#/components/responses/ObjectResponse' }
  /api/v1/work-packages/{workPackageId}/versions/{versionId}/validate:
    post:
      tags: [Work Package]
      operationId: validateWorkPackageVersion
      summary: Validate a draft before publication
      parameters:
        - $ref: '#/components/parameters/WorkPackageId'
        - $ref: '#/components/parameters/VersionId'
      responses:
        '200': { $ref: '#/components/responses/ObjectResponse' }
  /api/v1/work-packages/{workPackageId}/versions/{versionId}/publish:
    post:
      tags: [Work Package]
      operationId: publishWorkPackageVersion
      summary: Publish a validated work package version
      parameters:
        - $ref: '#/components/parameters/WorkPackageId'
        - $ref: '#/components/parameters/VersionId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PublishEffectiveWindow' }
      responses:
        '200': { $ref: '#/components/responses/ObjectResponse' }
  /api/v1/work-packages/{workPackageId}/versions/{versionId}/retire:
    post:
      tags: [Work Package]
      operationId: retireWorkPackageVersion
      summary: Retire a published work package version
      parameters:
        - $ref: '#/components/parameters/WorkPackageId'
        - $ref: '#/components/parameters/VersionId'
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RetireWorkPackageVersion' }
      responses:
        '200': { $ref: '#/components/responses/ObjectResponse' }
  /api/v1/work-packages/{workPackageId}/allocations:
    get:
      tags: [Work Package]
      operationId: listWorkPackageAllocations
      summary: List work package allocations
      parameters:
        - $ref: '#/components/parameters/WorkPackageId'
      responses:
        '200': { $ref: '#/components/responses/ListResponse' }
    post:
      tags: [Work Package]
      operationId: createWorkPackageAllocation
      summary: Allocate a published work package version to an assignment and target organization
      parameters:
        - $ref: '#/components/parameters/WorkPackageId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateWorkPackageAllocation' }
      responses:
        '201': { $ref: '#/components/responses/ObjectResponse' }
...
  /api/v1/work-expectations:
    get:
      tags: [Work Package]
      operationId: listWorkExpectations
      summary: List visible generated work expectations
      parameters:
        - { name: status, in: query, schema: { type: string } }
        - { name: positionAssignmentId, in: query, schema: { type: string, format: uuid } }
        - { name: targetOrgUnitId, in: query, schema: { type: string, format: uuid } }
        - { name: businessDate, in: query, schema: { type: string, format: date } }
      responses:
        '200': { $ref: '#/components/responses/ListResponse' }
  /api/v1/my/work-expectations:
    get:
      tags: [Work Package]
      operationId: listMyWorkExpectations
      summary: List expectations assigned to the current actor's active position assignments
      parameters:
        - { name: status, in: query, schema: { type: string } }
        - { name: businessDate, in: query, schema: { type: string, format: date } }
      responses:
        '200': { $ref: '#/components/responses/ListResponse' }
  /api/v1/team/work-expectations:
    get:
      tags: [Work Package]
      operationId: listTeamWorkExpectations
      summary: List expectations for the current manager's team
      parameters:
        - { name: status, in: query, schema: { type: string } }
        - { name: businessDate, in: query, schema: { type: string, format: date } }
      responses:
        '200': { $ref: '#/components/responses/ListResponse' }
  /api/v1/work-expectations/{expectationId}:
    get:
      tags: [Work Package]
      operationId: getWorkExpectation
      summary: Get an expectation and its immutable source snapshot
      parameters:
        - $ref: '#/components/parameters/ExpectationId'
      responses:
        '200': { $ref: '#/components/responses/ObjectResponse' }
  /api/v1/work-expectations/actions/generate:
    post:
      tags: [Work Package]
      operationId: generateWorkExpectations
      summary: Generate idempotent expectations for an assignment, organization and business date
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GenerateWorkExpectations' }
      responses:
        '201': { $ref: '#/components/responses/ObjectResponse' }
  /api/v1/work-expectations/sla/process:
    post:
      tags: [Work Package]
      operationId: processWorkExpectationSla
      summary: Mark overdue work expectations MISSED and emit WorkExpectationMissed events
      description: >-
        Requires work-package.manage. Candidates in PLANNED, AVAILABLE or IN_PROGRESS
        whose dueAt is before the processing time are changed to MISSED. Each changed
        expectation produces an audit record and transactional outbox event. Processing
        is tenant-isolated, batched, ordered and safe under concurrent calls.
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 500, default: 100 }
      responses:
        '200':
          description: SLA processing result
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WorkExpectationSlaProcessResult' }
        '400': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/work-expectations/{expectationId}/waive:
    post:
      tags: [Work Package]
      operationId: waiveWorkExpectation
      summary: Waive an expectation using optimistic concurrency
      parameters:
        - $ref: '#/components/parameters/ExpectationId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ExpectationAction' }
      responses:
        '200': { $ref: '#/components/responses/ObjectResponse' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/work-expectations/{expectationId}/cancel:
    post:
      tags: [Work Package]
      operationId: cancelWorkExpectation
      summary: Cancel an expectation using optimistic concurrency
      parameters:
        - $ref: '#/components/parameters/ExpectationId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ExpectationAction' }
      responses:
        '200': { $ref: '#/components/responses/ObjectResponse' }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Trusted production JWT or SSO access token.
    TenantHeader:
      type: apiKey
      in: header
      name: X-Tenant-Id
      description: Local integration only; requires DEV_HEADER_AUTH_ENABLED=true.
    ActorHeader:
      type: apiKey
      in: header
      name: X-Actor-Id
      description: Local integration only; requires DEV_HEADER_AUTH_ENABLED=true.
  parameters:
...
    VersionId: { name: versionId, in: path, required: true, schema: { type: string, format: uuid } }
...
    WorkPackageId: { name: workPackageId, in: path, required: true, schema: { type: string, format: uuid } }
    ExpectationId: { name: expectationId, in: path, required: true, schema: { type: string, format: uuid } }
...
  responses:
    ListResponse:
      description: List response
      content:
        application/json:
          schema:
            type: array
            items: { type: object, additionalProperties: true }
    ObjectResponse:
      description: Object response
      content:
        application/json:
          schema: { type: object, additionalProperties: true }
    Problem:
      description: RFC 9457 problem detail
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    Forbidden:
      description: Permission denied or organization scope violation
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Conflict:
      description: Optimistic version or idempotency conflict
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Problem:
      type: object
      properties:
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
...
    ApplicabilityScope:
      type: object
      required: [scopeType]
      properties:
        scopeType: { type: string, enum: [TENANT, BRAND, ORG_UNIT, ORG_TREE, POSITION] }
        brandId: { type: [string, 'null'], format: uuid }
        orgUnitId: { type: [string, 'null'], format: uuid }
        positionId: { type: [string, 'null'], format: uuid }
...
    PublishEffectiveWindow:
      type: object
      required: [effectiveFrom]
      properties:
        effectiveFrom: { type: string, format: date-time }
        effectiveTo: { type: [string, 'null'], format: date-time }
    CreateWorkPackage:
      type: object
      required: [code, name, positionId]
      properties:
        code: { type: string }
        name: { type: string }
        description: { type: [string, 'null'] }
        positionId: { type: string, format: uuid }
        ownerOrgUnitId: { type: [string, 'null'], format: uuid }
    CreateWorkPackageVersion:
      type: object
      required: [title]
      properties:
        title: { type: string }
        description: { type: [string, 'null'] }
    WorkPackageStandardLink:
      type: object
      required: [standardVersionId, usageType]
      properties:
        standardVersionId: { type: string, format: uuid }
        usageType: { type: string }
        weight: { type: [number, 'null'] }
    WorkPackageResponsibility:
      type: object
      required: [participantType, resolverType]
      properties:
        participantType: { type: string }
        resolverType: { type: string }
        positionId: { type: [string, 'null'], format: uuid }
        scopeStrategy: { type: [string, 'null'] }
        escalationLevel: { type: [integer, 'null'], minimum: 0 }
    WorkPackageItem:
      type: object
      required: [itemCode, name, itemType, formVersionId, periodType, responsibilities]
      properties:
        itemCode: { type: string }
        name: { type: string }
        description: { type: [string, 'null'] }
        itemType: { type: string }
        formVersionId: { type: string, format: uuid }
        sortOrder: { type: [integer, 'null'] }
        required: { type: [boolean, 'null'] }
        periodType: { type: string }
        timezoneMode: { type: [string, 'null'] }
        fixedTimezone: { type: [string, 'null'] }
        workWindowStart: { type: [string, 'null'], format: time }
        workWindowEnd: { type: [string, 'null'], format: time }
        dueLocalTime: { type: [string, 'null'], format: time }
        graceMinutes: { type: [integer, 'null'], minimum: 0 }
        weekdays: { type: [array, 'null'], items: { type: integer, minimum: 1, maximum: 7 } }
        dayOfMonth: { type: [integer, 'null'], minimum: 1, maximum: 31 }
        holidayPolicy: { type: [string, 'null'] }
        waiverAllowed: { type: [boolean, 'null'] }
        targetGranularity: { type: [string, 'null'] }
        reviewMode: { type: [string, 'null'] }
        standards: { type: [array, 'null'], items: { $ref: '#/components/schemas/WorkPackageStandardLink' } }
        responsibilities:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/WorkPackageResponsibility' }
        submissionPolicy: { $ref: '#/components/schemas/WorkSubmissionPolicy' }
    WorkSubmissionPolicy:
      type: object
      properties:
        completionStatementRequired: { type: boolean, default: true }
        exceptionStatementRequired: { type: boolean, default: false }
        nextActionRequired: { type: boolean, default: false }
        attachmentRequired: { type: boolean, default: false }
        maxAttachments: { type: integer, minimum: 0, maximum: 10, default: 10 }
        maxFileSizeBytes: { type: integer, minimum: 1, maximum: 20971520, default: 20971520 }
        allowedExtensions:
          type: array
          items: { type: string, enum: [jpg, jpeg, png, pdf, docx, xlsx] }
    UpdateWorkPackageVersion:
      type: object
      required: [title, scopes, items]
      properties:
        title: { type: string }
        description: { type: [string, 'null'] }
        scopes: { type: array, minItems: 1, items: { $ref: '#/components/schemas/ApplicabilityScope' } }
        items: { type: array, minItems: 1, items: { $ref: '#/components/schemas/WorkPackageItem' } }
    RetireWorkPackageVersion:
      type: object
      properties:
        effectiveTo: { type: [string, 'null'], format: date-time }
    CreateWorkPackageAllocation:
      type: object
      required: [workPackageVersionId, positionAssignmentId, targetOrgUnitId, validFrom]
      properties:
        workPackageVersionId: { type: string, format: uuid }
        positionAssignmentId: { type: string, format: uuid }
        targetOrgUnitId: { type: string, format: uuid }
        validFrom: { type: string, format: date }
        validTo: { type: [string, 'null'], format: date }
        allocationSource: { type: [string, 'null'] }
...
    GenerateWorkExpectations:
      type: object
      required: [positionAssignmentId, targetOrgUnitId, businessDate, periodType]
      properties:
        positionAssignmentId: { type: string, format: uuid }
        targetOrgUnitId: { type: string, format: uuid }
        businessDate: { type: string, format: date }
        periodType: { type: string }
        dutyPeriodId: { type: [string, 'null'], format: uuid }
    WorkExpectationSlaProcessResult:
      type: object
      required: [processedCount, batchLimit, expectationIds, processedAt]
      properties:
        processedCount:
          type: integer
          minimum: 0
          description: Expectations changed to MISSED in this pass.
        batchLimit:
          type: integer
          minimum: 1
          maximum: 500
          description: Validated maximum candidate count for this pass.
        expectationIds:
          type: array
          description: Ordered IDs changed to MISSED and emitted to the outbox.
          items: { type: string, format: uuid }
        processedAt:
          type: string
          format: date-time
          description: UTC completion timestamp for this pass.
    ExpectationAction:
      type: object
      required: [reason, expectedVersion]
      properties:
        reason: { type: string }
        expectedVersion: { type: integer, format: int64, minimum: 0 }
...